Cleanup `engine` option and remove trailing spaces.
The `ApolloServer` constructor options allow `engine` to be set to
`false`.  While the alternative of setting it to `undefined` (as it was
before) does effectively work the same, having the explicit `undefined` in
there gave the impression that `undefined` might be a desired attribute for
the `engine` option.  Plus, this was a bit more concise.

Also, removed some trailing spaces.  All in all this was mostly nitpicks.